Adding Wii Camera Libary and some user code
aaWiiCamera is the library for the camera (its called aaWii... so that
is show up at the top of the list no other reason). there are notes in
aaWiiCamera.cpp that denote what work needs to be done to get the i2c
running. This was all compiling on my machine at time of upload.